% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoUpdate.R
\name{autoInitConv}
\alias{autoInitConv}
\title{Run MCMC iterations until initial convergence}
\usage{
autoInitConv(x, iter = 5000, maxit = 95000, Rhat = 1.1)
}
\arguments{
\item{x}{An object of class `FluHMM', from which to generate posterior samples.}

\item{iter}{Number of iterations to run.}

\item{maxit}{Maximum number of iterations to run before giving up.}

\item{Rhat}{Gelman-Rubin diagnostic cutoff value to determine convergence
(only for the sigma[1] parameter).}
}
\value{
None. The function mutates its argument `x' directly.
}
\description{
This function is not normally called by the user; it is called from \code{\link{FluHMM}} provided
\code{initConv=TRUE} (the default). It generates posterior samples from the model repeatedly
until convergence is reached for the sigma[1] parameter (this is called "initial convergence").
}
\details{
The sigma[1] parameter (standard deviation of the pre-epidemic phase) is of primary
   importance in the model, since the pre-epidemic phase comes first and its correct identification
   is the basis on which to estimate the other phases. If convergence for sigma[1] has been reached,
   the other parameters in the model are very likely to have converged too, with the exception of
   beta[2] and beta[3] (slopes of the epidemic growth and epidemic decline phase); the latter mix
   more slowly and may necessitate a longer, and possibly thinned, posterior sample.

   Therefore "initial convergence" is defined as convergence for the sigma[1] parameter. Unless
   this is achieved, it is inadvisable to use the posterior samples for any inference at all. Only
   _after_ this has been achieved can a new posterior sample be generated (using \code{\link{update}}).
   Then convergence for all parameters is checked again and, if not achieved, a new sample is generated
   from scratch or the current one further is further extended.
}
