% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.mcmc.R
\name{c.mcmc}
\alias{c.mcmc}
\title{Combine an mcmc object}
\usage{
\method{c}{mcmc}(..., start = NA, thin = NA)
}
\arguments{
\item{...}{`mcmc' objects to be concatenated}

\item{start}{Iteration number of first observation in the chain, for the resulting object.
If \code{NA}, the first iteration number of the first `mcmc' object is used.}

\item{thin}{Thinning interval between consecutive observations, for the resulting object.
If \code{NA}, the thinning interval of the first `mcmc' object is used.}
}
\value{
Returns the combined `mcmc' object.
}
\description{
This method combines `mcmc' objects into a single `mcmc' object.
Objects must have come from the same model and have the same variable names.
}
