% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compress}
\alias{compress}
\title{Compress a FluHMM object}
\usage{
compress(x)
}
\arguments{
\item{x}{Object (of class `FluHMM') to be compressed.}
}
\value{
The function does NOT return anything, it mutates its argument directly.
}
\description{
This methods removes the MCMC samples and model object from the FluHMM object (elements
\code{cSample} and \code{model} of the FluHMM object), thus substantially reducing its size.
After this operation results are still available and the object can be plotted, but it is
not possible to generate further MCMC samples or run MCMC diagnostics. Use this function
to compress the object for long-term storage, when you are satisfied with the output.
}

