% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe.FluHMM}
\alias{describe.FluHMM}
\alias{describe}
\title{Describe a `FluHMM' object}
\usage{
\method{describe}{FluHMM}(x, recalc = FALSE)
}
\arguments{
\item{x}{An object of class `FluHMM', to be described.}

\item{recalc}{If \code{TRUE}, recalculate the most likely first epidemic weeks and most
likely peak intensity weeks, and store them in the object.}
}
\value{
None. The information are printed on the screen, and if \code{recalc=TRUE} they
   are also stored in the \code{descr} element of the FluHMM object. Thus the method
   mutates its argument `x' directly.
}
\description{
This method prints various interesting information about the fitted FluHMM object: the
number of initial weeks fixed to the pre-epidemic phase (K), the probability that the
epidemic wave has begun, the three most likely first epidemic weeks, the probability that
peak intensity has been reached, and the three most likely peak intensity weeks.
}
\details{
The current probability that the epidemic wave has begun is equal to the sum of the
   probabilities of phases 2 to 5 for the current (i.e. latest) week. Similarly, the current
   probability that peak intensity has been reached is the sum of the probabilities of phases
   3 to 5 for the current week.

   The most likely first epidemic week is calculated by iterating over the MCMC chains and
   finding, for each iteration, the first week where Z=2 (first week of the epidemic growth
   phase). Similarly, the most likely peak intensity week is calculated by iterating and
   finding the latest week where Z=2.
}
