% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePriors.R
\name{makePriors}
\alias{makePriors}
\title{Create informative prior from other FluHMM objects}
\usage{
makePriors(..., par = c("beta[1]", "beta[4]", "binc", "bprop"))
}
\arguments{
\item{...}{`FluHMM' objects to be used}

\item{par}{Which parameters to create priors for? A character vector.}
}
\value{
Returns a named list with the prior distribution parameters for 
each model parameter.
}
\description{
This function uses past end-of-season FluHMM objects and constructs
informative priors for the \code{beta[1]}, \code{binc}, \code{bprop} 
and \code{sigma[1]} model parameters.
}
