% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FluJointHMM.R
\name{plot.FluJointHMM}
\alias{plot.FluJointHMM}
\title{Plot a FluJointHMM object}
\usage{
\method{plot}{FluJointHMM}(x, xlab = "Week", ylab = "ILI rate",
  ylabIsol = "Influenza(+) samples", main = NA, xaxis = NA,
  showPs = TRUE, cexPs = 0.7, yexpand = 0.3, col = "red",
  mucol = "limegreen", hues = c(4, 0, 2, 5, 3)/6, rainbow = FALSE,
  ci = TRUE, alpha = 0.1)
}
\arguments{
\item{x}{An object of class `FluJointHMM' to be plotted.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis of the ILI/ARI rate sub-plot.}

\item{ylabIsol}{Label for the y-axis of the isolates sub-plot.}

\item{main}{Main label of the plot.}

\item{xaxis}{How to annotate the x-axis of the plot. If \code{NULL}, no axis is being drawn
(equivalent to \code{xaxt='n'}). If \code{xaxis=0}, the axis names are taken from the
\code{names} attribute of the \code{seasonRates} element of the FluHMM object. If
\code{xaxis=NA}, the \code{seasonRates} names if treated as strings of format YYYYWW, and
the week number is plotted. Otherwise, if a vector of length>1 is supplied, it is treated
as the week numbers to be plotted at the x-axis.}

\item{showPs}{If \code{TRUE} (the default), show the weekly posterior probabilities of each phase
as a series of numbers too, instead of only a color representation.}

\item{cexPs}{Character expansion factor (cex), i.e. size for plotting the posterior probabilities}

\item{yexpand}{Factor (as a percentage of the original y-axis size) to expand the y-axis, so that the
numbers of the posterior probabilities do not overlap with the rest of the plot.}

\item{col}{Color for plotting the weekly ILI/ARI rates}

\item{mucol}{Color for plotting the fitted weekly mean rates.}

\item{hues}{A numeric vector of length 5, with values between 0 and 1, containing the hue values
for each of the five phases in the model.}

\item{rainbow}{If \code{TRUE}, the mean rates and mean number of isolates for each of the
six chains are individually plotted as well.}

\item{ci}{If \code{TRUE}, semi-transparent 95% confidence bands are plotted around the rates,
provided the `FluJointHMM' object includes a logSE element (the log Standard Errors of the rates).}

\item{alpha}{Alpha transparency value for the confidence bands (a number between 0 and 1).}
}
\value{
None
}
\description{
This function plots the ILI/ARI rates and number of influenza-positive isolates stored
in a FluJointHMM object, superimposing the model results. The input and output is
mostly the same as \code{\link{plot.FluHMM}}, with an additional smaller sub-plot
of the isolates at the bottom.
}
\details{
This function plots the ILI/ARI rates with the week number on the x-axis. If the FluJointHMM
   object has a \code{seasonRates} element, this is plotted as a thin grey line and the \code{rates}
   are overlaid as a thick line with points, of color \code{col}.

   A similar subplot (a bar plot) of the influenza-positive isolates is drawn under the ILI/ARI rates plot.

   The posterior probabilities of the five epidemic phases per week (pre-epidemic, epidemic growth,
   epidemic plateau, epidemic decline and post-epidemic) are displayed with colored bars on the top
   of the plot and optionally as a vertical stack of numbers (if \code{showPs=TRUE}). The hue of
   each phase in the color representation is given in the \code{hues} argument, and the saturation
   is dependent on the posterior probability of each phase. If the vertical space of the plot is
   insufficient (especially if \code{showPs=TRUE}), expand it as necessary by increasing the
   \code{yexpand} argument.

   The mean weekly rates fitted by the model is plotted as a thick dotted line. If \code{rainbow=TRUE},
   the mean rates for each MCMC chain are plotted individually as well. These should be very close to
   each other; if very different, then the chains have probably not converged enough.

   In any case, if full convergence (for all model parameters) has not been reached, a clear warning
   will be displayed on the plot.
}
